Testing and confirming the service works.
