--- rl_infra/edge/__init__.py
@@ -1,3 +1,0 @@
-import config
-import robot_service
-import utils
