+++ rl_infra/edge/config.py
@@ -1,5 +1,6 @@
 DIST_PATH = "/distance"
 IMG_PATH = "/camera"
 MOVE_PATH = "/action"
-SERVER_HOST = "localhost"
+SERVER_HOST = "0.0.0.0"
+CLIENT_HOST = "dex.local"
 SERVER_PORT = 8080
