+++ rl_infra/online/utils/robot_client.py
@@ -1,0 +1,42 @@
+import json
+from typing import Tuple
+from PIL import Image
+
+from IPython import embed
+import numpy
+import requests
+
+from rl_infra.edge import config
+from rl_infra.edge.utils import uncompress_nparr
+
+
+class RobotClient:
+    url = f"http://{config.CLIENT_HOST}:{config.SERVER_PORT}"
+
+    @staticmethod
+    def sendAction(action: int, arg: int) -> None:
+        data = {"action": action, "arg": arg}
+        requests.post(
+            url=RobotClient.url + config.MOVE_PATH,
+            data=json.dumps(data),
+            headers={"Content-type": "application/json"},
+        )
+
+    @staticmethod
+    def getSensorReading() -> Tuple[numpy.ndarray, int]:
+        imgResponse = requests.get(
+            url=RobotClient.url + config.IMG_PATH,
+            headers={"Content-Type": "application/octet-stream"},
+        )
+        img = uncompress_nparr(imgResponse.content)
+
+        distResponse = requests.get(url=RobotClient.url + config.DIST_PATH)
+        dist = int(distResponse.content)
+
+        return img, dist
+
+    # This should go with offline data collection, but I am saving here for now.
+    @staticmethod
+    def saveArrayAsJpeg(img: numpy.ndarray, filePath: str) -> None:
+        im = Image.fromarray(img)
+        im.save(filePath)
