--- rl_infra/online/utils/robot_driver.py
@@ -1,26 +1,0 @@
-import requests
-
-from rl_infra.edge import config
-from rl_infra.edge.utils import uncompress_nparr
-from rl_infra.online.impl import RobotAction, RobotState
-
-
-class RobotClient:
-    url = f"http://{config.SERVER_HOST}:{config.SERVER_PORT}"
-
-    @staticmethod
-    def sendAction(action: RobotAction):
-        _ = requests.post(url=RobotClient.url + config.MOVE_PATH, data=str(action))
-
-    @staticmethod
-    def getSensorReading() -> RobotState:
-        imgResponse = requests.get(
-            url=RobotClient.url + config.IMG_PATH,
-            headers={"Content-Type": "application/octet-stream"},
-        )
-        img = uncompress_nparr(imgResponse.content)
-
-        distResponse = requests.get(url=RobotClient.url + config.DIST_PATH)
-        dist = int(distResponse.content)
-
-        return RobotState(img, dist)
